+++ src/projectile.cc
@@ -11,7 +11,7 @@ Projectile::Projectile(const std::string &type) :
 Projectile::Projectile(const std::string &type, Deserializer &deser) : Mob(type, deser) {
 }
 
-void Projectile::Start(RunningState &state, size_t id) {
+void Projectile::Start(RunningState &state, uint32_t id) {
   Log("Projectile::Start()\n");
   Mob::Start(state, id);
