+++ src/util.h
@@ -19,7 +19,7 @@ time_t getFileChangeTime(const std::string &name);
 float Wave(float x, float z, float t, float a = 0.2);
 
 std::vector<std::string> Tokenize(const char *line);
-size_t ParseSidesMask(const std::string &str);
+uint32_t ParseSidesMask(const std::string &str);
 
 int saveImage(const std::string &fileName, size_t w, size_t h, const uint8_t *rgb);
